♻️  refactor: [#55] Header 분리 - Public,Private
