+++ src/components/layout/Header.tsx
@@ -1,58 +1,8 @@
-import { Link, useNavigate } from 'react-router-dom';
-import { useRecoilState } from 'recoil';
 import styled from 'styled-components';
-
-import imgUrl from '@assets/logo.png';
-import { IoNotifications, IoPerson, IoLogOutOutline } from 'react-icons/io5';
+import { Link } from 'react-router-dom';
 import { Button } from '@components/common';
-import { PAGE_URL } from '@constants';
-import { authService } from '@apis';
-import { loginState } from '@state/atom';
 
-export default function Header() {
-  const [isLoggedIn, setIsLoggedIn] = useRecoilState(loginState);
-
-  const navigate = useNavigate();
-  const handleLogout = () => {
-    authService.signOut();
-    setIsLoggedIn(false);
-    navigate(PAGE_URL.ROOT);
-  };
-
-  return (
-    <SHeader>
-      <Logo to={PAGE_URL.ROOT}>
-        <img src={imgUrl} alt="logo_icon" />
-      </Logo>
-
-      {isLoggedIn && (
-        <SMenu>
-          <IoNotifications />
-          <SLink to={PAGE_URL.MYPAGE}>
-            <IoPerson />
-          </SLink>
-          <IoLogOutOutline onClick={handleLogout} />
-        </SMenu>
-      )}
-
-      {!isLoggedIn && (
-        <SMenu>
-          <SLink to={PAGE_URL.SIGN_IN}>
-            <HeaderButton type="button" styleType="neutral" size="small">
-              로그인
-            </HeaderButton>
-          </SLink>
-          <SLink to={PAGE_URL.SIGN_UP}>
-            <HeaderButton type="button" styleType="solidPositive" size="small">
-              회원가입
-            </HeaderButton>
-          </SLink>
-        </SMenu>
-      )}
-    </SHeader>
-  );
-}
-const SHeader = styled.header`
+export const Header = styled.header`
   position: sticky;
   top: 0;
   z-index: 1;
@@ -66,15 +16,7 @@ const SHeader = styled.header`
   min-width: 375px;
 `;
 
-const Logo = styled(Link)`
-  height: 100%;
-  cursor: pointer;
-  img {
-    height: 3.75rem;
-  }
-`;
-
-const SMenu = styled.div`
+export const Menu = styled.div`
   display: flex;
   svg {
     cursor: pointer;
@@ -85,10 +27,10 @@ const SMenu = styled.div`
   }
 `;
 
-const SLink = styled(Link)`
+export const MenuLink = styled(Link)`
   text-decoration: none;
 `;
 
-const HeaderButton = styled(Button)`
+export const HeaderButton = styled(Button)`
   margin: 0 0.4rem;
 `;
